+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [ADD NEW VERSION HERE] - ADD DATE HERE
 ### Fixed
+ - Fix incorrect 'wult calc' warning about non-numeric metrics.
 ### Added
 ### Removed
 ### Changed